removed old react, added tests
