--- package/client/index.ts
@@ -1,3 +1,0 @@
-import { NeuronClient } from "./NeuronClient";
-
-export { NeuronClient };
