--- tutorials/react/Counter-New/App.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import Counter from "./Counter";
-
-export default function App() {
-  return (
-    <>
-      <div style={{ height: "75vh" }}>
-        <Counter />
-      </div>
-    </>
-  );
-}
